requirements.txt
fastapi
sqlalchemy
uvicorn[standard]
psycopg2-binary
python-jose
bcrypt
passlib
